chore: update script for starting the dev environment

+++ scripts/start_development_environment.sh
@@ -2,6 +2,11 @@
 
 set -euo pipefail
 
+if [[ ! -f client_secrets.json ]]; then
+    echo "client_secrets.json not found. Please create one and put it in the project root directory."
+    exit 1
+fi
+
 COMPOSE_PROFILES=dev
 
 docker compose down
